+++ R/station_name.R
@@ -1,4 +1,12 @@
+#' Retrieve Information about a Monitoring Station
+#'
+#' @param code Water monitoring station code (e.g. "FBLW")
 #' @export
+#' @examples
+#' library(acewater)
+#'
+#' station_name("FBLW")
+#'
 station_name <- function(code) {
-    purrr:::map_chr(code, { acewater:::stations[[.]]$name })
+    purrr::map_chr(code, function(x) stations[[x]]$name)
 }